Rename datahub.oauth.admin_sso app
This renames the `datahub.oauth.admin` Django app to `datahub.oauth.admin_sso` as it was conflicting with the `datahub.oauth.admin` module.
