+++ datahub/oauth/admin_sso/apps.py
@@ -4,4 +4,4 @@ from django.contrib.admin.apps import AdminConfig
 class OAuthAdminConfig(AdminConfig):
     """Required to override stock Django admin app with OAuth2 admin app."""
 
-    default_site = 'datahub.oauth.admin.admin.OAuth2AdminSite'
+    default_site = 'datahub.oauth.admin_sso.admin.OAuth2AdminSite'
