+++ datahub/oauth/admin_sso/test/test_admin.py
@@ -12,7 +12,7 @@ pytestmark = pytest.mark.django_db
 class TestAdminSite(AdminTestMixin):
     """Test that admin site login and logout views have been replaced with OAuth2 equivalents."""
 
-    @patch('datahub.oauth.admin.admin.login')
+    @patch('datahub.oauth.admin_sso.admin.login')
     def test_oauth2_login_view_is_called_when_requesting_admin_login(self, _login):
         """Tests that OAuth2 login view is called when accessing admin login."""
         _login.return_value = HttpResponse()
@@ -22,7 +22,7 @@ class TestAdminSite(AdminTestMixin):
 
         _login.assert_called_once()
 
-    @patch('datahub.oauth.admin.admin.logout')
+    @patch('datahub.oauth.admin_sso.admin.logout')
     def test_oauth2_login_view_is_called_when_requesting_admin_logout(self, _logout):
         """Tests that OAuth2 logout view is called when accessing admin logout."""
         _logout.return_value = HttpResponse()
